SpawnManager spawns in enemy customers
Enemy customers now get stunned when food is thrown at them. They spawn in within the area at random 15 to 30 second intervals and are never destroyed until the game ends. Temporarily, the scene restarts if the player is caught but the player cannot be caught if the enemy customer is stunned. Animations have been added as required.